third_party/blink/web_tests/fast/multicol/cssom-view-expected.txt

elm1


elm2
elm3
elm4
elm5

PASS mc.offsetLeft is 10
PASS mc.offsetTop is 15
PASS elm1.offsetLeft is mc.offsetLeft
PASS elm1.offsetTop is mc.offsetTop
PASS elm2.offsetLeft is > elm1.offsetLeft
PASS elm2.offsetTop is elm1.offsetTop
PASS elm3.offsetLeft is elm2.offsetLeft
PASS elm3.offsetTop is > elm2.offsetTop
PASS elm4.offsetLeft is > elm3.offsetLeft
PASS elm4.offsetTop is elm1.offsetTop
PASS elm5.offsetLeft is elm4.offsetLeft
PASS elm5.offsetTop is elm3.offsetTop
PASS rects.length is 1
PASS rects[0].left is elements[i].offsetLeft
PASS rects[0].top is elements[i].offsetTop
PASS rects.length is 1
PASS rects[0].left is elements[i].offsetLeft
PASS rects[0].top is elements[i].offsetTop
PASS rects.length is 1
PASS rects[0].left is elements[i].offsetLeft
PASS rects[0].top is elements[i].offsetTop
PASS rects.length is 1
PASS rects[0].left is elements[i].offsetLeft
PASS rects[0].top is elements[i].offsetTop
PASS rects.length is 1
PASS rects[0].left is elements[i].offsetLeft
PASS rects[0].top is elements[i].offsetTop
PASS successfullyParsed is true

TEST COMPLETE
